Update unit tests that they work with Windows CPython version 3.7.13 (no problem on Linux).

Add conditions that the unit tests run with v221 version.

Black formatting to be flake8 compliant.

Related work items: #681810

+++ tests/testfiles/pythonPlugins/all_types/integral_types_op.py
@@ -100,7 +100,7 @@ class SetOutVecDoubleOperator(CustomOperatorBase):
 class SetOutNpArrayIntOperator(CustomOperatorBase):
     def run(self):
         import numpy
-        out = numpy.ones((100,2), dtype=numpy.int)
+        out = numpy.ones((100,2), dtype=numpy.int32)
         self.set_output(0, out)
         self.set_succeeded()
